requirements.txt
streamlit>=1.31
pillow>=9.4
numpy>=1.23
opencv-python-headless>=4.8
